# frontend/src/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ApeRAG API
  description: ApeRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /marketplace/collections:
    get:
      summary: List all published Collections in marketplace
      description: Returns all Collections that are published to marketplace with current user's subscription status
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
      responses:
        '200':
          description: List of published Collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedCollectionList'
        '401':
          $ref: '#/components/responses/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/subscriptions:
    get:
      summary: Get user's subscribed Collections
      description: Returns all Collections that current user has subscribed to
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
      responses:
        '200':
          description: List of subscribed Collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedCollectionList'
        '401':
          $ref: '#/components/responses/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/subscribe:
    post:
      summary: Subscribe to a Collection
      description: Subscribe to a published Collection
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Successfully subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedCollection'
        '400':
          description: Bad request (e.g., Collection not published, self-subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '401':
          $ref: '#/components/responses/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '409':
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
    delete:
      summary: Unsubscribe from a Collection
      description: Remove subscription to a Collection
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '204':
          description: Successfully unsubscribed
        '401':
          $ref: '#/components/responses/failResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}:
    get:
      summary: Get MarketplaceCollection details (read-only)
      description: Get details of a subscribed Collection (read-only access)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: MarketplaceCollection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedCollection'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/documents:
    get:
      summary: List documents in MarketplaceCollection (read-only)
      description: Get document list for a subscribed Collection (read-only mode)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Search keyword
          schema:
            type: string
        - name: file_type
          in: query
          description: File type filter
          schema:
            type: string
      responses:
        '200':
          description: Document list (read-only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/documentList'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/documents/{document_id}/preview:
    get:
      summary: Preview document in MarketplaceCollection (read-only)
      description: Get document preview for a subscribed Collection (read-only mode)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      responses:
        '200':
          description: Document preview (read-only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/documentPreview'
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
  /marketplace/collections/{collection_id}/graph:
    get:
      summary: Get knowledge graph for MarketplaceCollection (read-only)
      description: Get knowledge graph data for a subscribed Collection (read-only mode)
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: node_limit
          in: query
          description: Maximum number of nodes
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: depth
          in: query
          description: Graph traversal depth
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 2
      responses:
        '200':
          description: Knowledge graph data (read-only)
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                  edges:
                    type: array
                    items:
                      type: object
                  collection_id:
                    type: string
                  read_only:
                    type: boolean
        '401':
          $ref: '#/components/responses/failResponse'
        '403':
          description: Access denied (need subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          $ref: '#/components/responses/failResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    pageResult:
      description: PageResult info
      properties:
        page_number:
          type: integer
          description: The page number
        page_size:
          type: integer
          description: The page size
        count:
          type: integer
          description: The total count of items
      type: object
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - RUNNING
            - COMPLETE
            - FAILED
            - DELETING
            - DELETED
        vector_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        fulltext_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        graph_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        summary_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vision_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vector_index_updated:
          type: string
          format: date-time
          description: Vector index last updated time
        fulltext_index_updated:
          type: string
          format: date-time
          description: Fulltext index last updated time
        graph_index_updated:
          type: string
          format: date-time
          description: Graph index last updated time
        summary_index_updated:
          type: string
          format: date-time
          description: Summary index last updated time
        vision_index_updated:
          type: string
          format: date-time
          description: Vision index last updated time
        summary:
          type: string
          description: Summary of the document
        config:
          type: string
        size:
          type: number
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    documentList:
      type: object
      description: A list of documents
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/document'
        pageResult:
          $ref: '#/components/schemas/pageResult'
...
    visionChunk:
      type: object
      properties:
        id:
          type: string
        asset_id:
          type: string
        text:
          type: string
        metadata:
          type: object
    documentPreview:
      type: object
      properties:
        doc_object_path:
          type: string
          nullable: true
          description: The path to the document object.
        doc_filename:
          type: string
          description: The name of the document.
        converted_pdf_object_path:
          type: string
          nullable: true
          description: The path to the converted PDF object.
        markdown_content:
          type: string
          description: The markdown content of the document.
        chunks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              metadata:
                type: object
        vision_chunks:
          type: array
          items:
            $ref: '#/components/schemas/visionChunk'
...
    SharedCollection:
      type: object
      description: Shared Collection information for marketplace users
      properties:
        id:
          type: string
          description: Collection ID
        title:
          type: string
          description: Collection title
        description:
          type: string
          description: Collection description
        owner_user_id:
          type: string
          description: Original owner user ID
        owner_username:
          type: string
          description: Original owner username
        subscription_id:
          type: string
          nullable: true
          description: Subscription record ID (has value if subscribed, null if not subscribed)
        gmt_subscribed:
          type: string
          format: date-time
          nullable: true
          description: Subscription time (only has value when subscribed)
      required:
        - id
        - title
        - description
        - owner_user_id
        - owner_username
    SharedCollectionList:
      type: object
      description: Shared Collection list response
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SharedCollection'
          description: List of shared Collections
        total:
          type: integer
          description: Total count (for pagination)
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Page size
      required:
        - items
        - total
        - page
        - page_size
...
  responses:
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request